test(datatable): refactor to use RTL
